+++ CHANGES.md
@@ -1,4 +1,12 @@
-## v1.0.3 (in development)
+## v1.0.4 (in development)
+
+### New Features
+
+
+### Fixes
+
+
+## v1.0.3
 
 ### New Features
 
@@ -6,8 +14,10 @@
   for e.g. testing purposes.
 - A new geoDB docker backend image has been added. This image can be used to set up a postgis database with a 
   pre-installed geodb database extension 
-- Added a new `count_by_bbox` filter function 
-
+- Added a new `count_by_bbox` filter function
+- Allowed to provide limit and offset in function `get_collection`
+- Added the new function `get_collection_bbox`, that returns the bounding box
+  of all geometries within a collection.
 
 ### Fixes
 
@@ -15,7 +25,9 @@
 - Improved heading in notebooks
 - Removed auto cp of notebooks in Makefile.
 - Ensured that the geoDB client works with Keycloak
-- Reduced substantially the number of times the token is received from auth service 
+- Reduced substantially the number of times the token is received from auth service
+- Fixed bug that caused unpublishing of collections with upper case letters 
+  in name to fail
 
 ## v1.0.2